style(GqlEditor): conform more with style rfc
