--- src/GraphQLEditor/actions.js
@@ -1,18 +1,0 @@
-import { graphqlSchemaUrl } from '../localconf';
-import { fetchWrapper } from '../actions';
-
-
-export const fetchSchema = () => dispatch => dispatch(fetchWrapper({
-  path: graphqlSchemaUrl,
-  handler: receiveSchema,
-}));
-
-export const receiveSchema = ({ status, data }) => {
-  switch (status) {
-  case 200:
-    return {
-      type: 'RECEIVE_SCHEMA_LOGIN',
-      schema: data,
-    };
-  }
-};
